+++ dataiter/data_frame.py
@@ -585,6 +585,7 @@ class DataFrame(dict):
             n = self[name].is_missing().sum()
             if n == 0: continue
             nas = nas.rbind(DataFrame(column=name, nna=n))
+        if not nas: return
         nas.pna = [f"{100*x/self.nrow:.1f}%" for x in nas.nna]
         nas.colnames = [x.upper() for x in nas.colnames]
         print(nas)