+++ NEWS.md
@@ -1,7 +1,8 @@
-PENDING: dataiter 0.24.1/0.25
-=============================
+PENDING: dataiter 0.25
+======================
 
 * `DataFrame.print_missing_counts`: Fix when nothing missing
+* `Vector.replace_missing`: New method
 
 2021-10-27: dataiter 0.24
 =========================